nu lyckas jag hämta och uppdatera state variable och frontend funkar typ
