+++ client/src/App.css
@@ -16,9 +16,8 @@ font-family: 'Krona One', sans-serif; */
   display: flex;
   margin: 0 auto;
   margin-left: -5px;
-
   flex-direction: column;
-  align-items: center; 
+  align-items: center;
 
 }
 
@@ -28,7 +27,7 @@ font-family: 'Krona One', sans-serif; */
   grid-template-rows: repeat(15, 1fr);
   grid-gap: 0px;
   max-width: 765px;
-  border: solid #967248 10px; 
+  border: solid #967248 10px;
   border-radius: 8px;
 }
 
@@ -37,9 +36,9 @@ font-family: 'Krona One', sans-serif; */
   height: 30px;
   border: rgb(0, 0, 0) 1.1px solid;
   background-color: whitesmoke;
-  background-color: FFFFFF;
-
+  /* background-color: FFFFFF; */
 }
+
 
 .b-circle {
   width: 100%;
@@ -62,25 +61,32 @@ font-family: 'Krona One', sans-serif; */
   flex-direction: column;
   align-items: center;
   position: relative;
-  margin-top: -29.7rem
+  margin-top: -29.3rem
 }
-
+/*
 .i-board {
-  display: grid;
-  grid-template-columns: repeat(16, 1fr);
-  grid-template-rows: repeat(16, 1fr);
-  grid-gap: 0px;
-  max-width: 765px;
-}
+  margin-top: -2px;
+  margin-left: 2px;
+  max-width: 470px;
+  max-height: 470px;
+} */
 
 .invisible-cell {
-  width: 30px;
-  height: 30px;
+  z-index: 10;
+  width: 27px;
+  height: 27px;
   border: white 1px solid;
   background-color: transparent;
   border-color: transparent;
-  z-index: 1;
+  padding: 2px;
 }
+
+.invisible-row {
+  display: flex;
+  flex-direction: row;
+  height: 30px;
+}
+
 
 /* Navbar */
 .navbar {
@@ -321,7 +327,7 @@ font-family: 'Krona One', sans-serif; */
   align-items: center;
   border-radius: 4px ;
 margin-bottom: 10%;
- 
+
 }
 .board-timer{
   display: flex;
@@ -334,7 +340,7 @@ margin-bottom: 10%;
   row-gap: 7px;
  margin: 20px 40px 40px 40px;
  }
- 
+
  .black-circle{
   width: 30px;
   height: 30px;
@@ -462,7 +468,7 @@ margin-bottom: 10%;
     line-height: 1.5;
     border-radius: 6px;
     box-shadow: 5px 5px 10px #00E0FF;
-   
+
 }
 
  .footer {
@@ -479,7 +485,7 @@ margin-bottom: 10%;
     text-align: center;
  }
 
- .losing-popup{ 
+ .losing-popup{
   width: 300px;
    height: 200px; border: 1px solid #fff;
    display: flex;
@@ -512,7 +518,7 @@ border: none;
 border-radius: 4px;
  }
  .losing-popup-content #close-popup{
- 
+
 border: none;
 background-color: transparent;
 
@@ -530,14 +536,14 @@ cursor: pointer;
 
  /* Responsive navbar, section, spelsida*/
 @media (min-width:375px) and (max-width:1000px){
-  
+
   body{
     margin: 0;
     padding: 0;
     overflow-y: scroll;
     overflow-x: hidden;
     width:100%;
-    
+
   }
   .logonav img{
     height: 55px;
@@ -547,7 +553,7 @@ cursor: pointer;
     display: flex;
     align-items: center;
     justify-content: space-between;
-    
+
   }
   .navigation-links {
   display: flex;
@@ -561,7 +567,7 @@ row-gap: 10px;
   }
   .contact-link{
     visibility: hidden;
-   
+
   }
   .navbar .navigation-links li {
     margin: 0 ;
@@ -589,7 +595,7 @@ row-gap: 10px;
   flex-direction: column;
   width: 100%;
   overflow-x: hidden;
- 
+
   row-gap: 60px;
   padding-bottom: 30%;
   }
@@ -598,7 +604,7 @@ row-gap: 10px;
     flex-direction: column;
     align-items: center;
     margin: 40px 0px 0px 0px;
-    
+
      }
   .board-container {
     display: flex;
@@ -606,11 +612,11 @@ row-gap: 10px;
     margin-left: -5px;
     flex-direction: column;
     align-items: center;
-   
-    margin:8px 0px 8px 0px; 
-   
+
+    margin:8px 0px 8px 0px;
+
   }
-  
+
   .board {
     display: grid;
     grid-template-columns: repeat(15, 1fr);
@@ -618,22 +624,21 @@ row-gap: 10px;
     grid-gap: 0px;
     max-width: 350px;
     overflow-x: hidden;
-border: none;
+    border: none;
     border-radius: 8px;
   }
- 
+
    .board-top{
     width: 348px;
    }
    .board-bottom{
     width: 348px;
-  
+
    }
    .footer{
     display: flex;
    justify-content: center;
 
    }
-  
-}
 
+}
