+++ client/src/api/Gamedata.js
@@ -15,9 +15,9 @@ const getGameData = async () => {
 const getGameBoard = async (boardId) => {
   try {
     const response = await axios.get(`/api/get_board/${boardId}`);
-   
+
     return response.data;
-   
+
   } catch (error) {
     console.error(error);
   }
@@ -52,7 +52,7 @@ const addPlayer2 = async (playerName) => {
     console.error(error);
   }
 };
- 
+
 
 
 const changeBoardState = async (boardId, newState) => {
@@ -65,8 +65,9 @@ const changeBoardState = async (boardId, newState) => {
   };
 
 const clickTile = async(rowId, colId, playerName, isPlayer1orPlayer2) => {
-  const url = new URL(window.location.href);
-  const boardId = url.pathname.split('/').pop();
+  // const url = new URL(window.location.href);
+  // const boardId = url.pathname.split('/').pop();
+  const boardId = "2ca69660-fa1b-45be-b53d-29ea6ca839f9";
   try {
     const response = await axios.post(`/api/click_tile/${boardId}`, { rowId, colId, playerName, isPlayer1orPlayer2 });
     return response.data;
@@ -77,4 +78,3 @@ const clickTile = async(rowId, colId, playerName, isPlayer1orPlayer2) => {
 
 
 export { getGameData, addPlayer1, addPlayer2, createNewGameBoard, clickTile, getGameBoard, changeBoardState };
-
